REF: Refactor SPEC-like scans to use planify.
